reorganisation of example code and demo house part I
