+++ examples/02_gpsdata/gpsfiles.py
