+++ examples/02_gpsdata/haversine.py
