+++ examples/03_gpsdata_oo/gpspoint.py
