+++ examples/03_gpsdata_oo/haversine.py
