+++ examples/03_gpsdata_oo/main.py
