+++ examples/04_bikecomputer_oomodel/Sykkelcomputer_UML_sketch.pdf
