+++ examples/gpsdata_reference/gpsanalysis.py
