+++ examples/gpsdata_reference/plots/heights.png
