+++ examples/gpsdata_reference/plots/speeds.png
