+++ exercises/exercise1.md
@@ -1,6 +1,6 @@
 # GPS Computer
 
-Åpne og kjøre filen [main.py](../examples/gpsdata/main.py) i  mappen `/examples/gpsdata/` for å få en ide om hva vi skal utvikle.
+Åpne og kjøre filen [main.py](../examples/02_gpsdata/main.py) i  mappen `/examples/gpsdata/` for å få en ide om hva vi skal utvikle.
 Som du ser når du velger en datasett så får du et resultat som ligner dette her:
 ```
 Selected dataset: short
@@ -11,4 +11,4 @@ Max speed      : NotImplemented km/t
 Average speed  : NotImplemented km/t
 ```
 Det betyr er at selve GSP kalkulatoren er ikke helt ferdig implementert.
-Din oppgave er å legge inn den manglende kode i de funksjonene i [gspcomputer.py](../examples/gpsdata/gpscomputer.py).
+Din oppgave er å legge inn den manglende kode i de funksjonene i [gspcomputer.py](../examples/02_gpsdata/gpscomputer.py).
