+++ project/demo.md
@@ -1,0 +1,18 @@
+# Demo-Hus
+
+For å teste din implementasjon på et konkret eksempel skal vi fra nå av baserer på et _demohus_.
+Du finner en plantegning av dette huset nedenfor:
+
+## Etasje 1
+
+![Etasjeplan Etasje 1](../resources/images/DemoHusFloorplan1.svg)
+
+## Etasje 2
+
+![Etasjeplane Etasje 2](../resources/images/DemoHusFloorplan2.svg)
+
+## Enheter
+
+På etasjeplanen ble også smarte enheter tegnet inn.
+Sensorer er tegnet i oransje og sensorer er tegnet i grønn.
+Tabellen nedenfor gir detaljer hva type sensor/aktuator skjuler seg bak de enkelte tallene.
