+++ weeks/3-intro-oo/week3.md
@@ -102,6 +102,6 @@ Etter forlesningen skulle du
   - konstruktor, og
 - vet hva ideen "Tell, don't ask!" betyr
 
-Du finner slides'ene om OO på Canvas og kodeeksempler fra forelesningen [her](../../examples/gpsdata/min_oo.py)
+Du finner slides'ene om OO på Canvas og kodeeksempler fra forelesningen:
 
-[Kode for det objekt-orienterte GPS data eksemplet](../../examples/gpsdata_oo/) 
+[Kode for det objekt-orienterte GPS data eksemplet](../../examples/03_gpsdata_oo/)
